requirements.txt
python-telegram-bot==21.10
playwright==1.42.0
playwright-stealth==1.0.6